requirements.txt
Django>=6.0.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
psycopg2-binary>=2.9    
python-decouple>=3.7
drf-yasg>=1.22  
pytest-django>=4.5                